Move Home Connect service actions to a actions.py
