+++ tests/components/home_connect/snapshots/test_actions.ambr
